Update spec guide pages with static_page olytics tracking
update the context that is being passed to GTM in order to track the spec guides as staic pages within olytics.  This will push the new static_page object
// at minimum
{
  title: “blah”
}

to the gtm data layer.  This is used in the olytics script to fire a static page view with { behaviorId, name, pageType }.  We can look at adding more as needed.