initial setup of localized default content/section routing/templates
Setup the basic dafault routing for section  & content with site localized version of basic routes to start with.  This will allow for the template overrides and adjustments needed for the IronPros site setup.

This a basic copy down with minor adjustments to accoutn for the lack of GAM.  We will continue to modifies these while using global components.